requirements.txt
aioredis==1.3.1
amqp==5.0.2
appdirs==1.4.4
asgiref==3.2.10
async-timeout==3.0.1
attrs==20.3.0
autobahn==20.12.3
Automat==20.2.0
autopep8==1.5.5
billiard==3.6.3.0
black==20.8b1
celery==5.0.5
certifi==2020.12.5
cffi==1.14.4
channels==2.4.0
channels-redis==3.2.0
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
cloudinary==1.24.0
constantly==15.1.0
coverage==5.5
cryptography==3.3.1
daphne==2.5.0
dj-database-url==0.5.0
Django==3.1.5
django-cachalot==2.3.3
django-debug-toolbar==3.2
# django-heroku==0.3.1
django-redis==4.12.1
djangorestframework==3.12.2
drf-extra-fields==3.0.4
fcm-django==0.3.7
flake8==3.8.4
gunicorn==20.0.4
h11==0.12.0
hiredis==1.1.0
httpcore==0.12.3
hyperlink==21.0.0
idna==2.10
importlib-metadata==3.3.0
incremental==17.5.0
iniconfig==1.1.1
kombu==5.0.2
mccabe==0.6.1
msgpack==1.0.2
mypy-extensions==0.4.3
packaging==20.9
pathspec==0.8.1
phonenumbers==8.12.15
Pillow==8.1.0
pluggy==0.13.1
prompt-toolkit==3.0.10
psutil==5.8.0
psycopg2-binary==2.8.6
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
pyfcm==1.4.9
pyflakes==2.2.0
PyHamcrest==2.0.2
PyJWT==2.0.0
pyOpenSSL==20.0.1
pyparsing==2.4.7
pytest==6.2.2
pytest-cov==2.11.1
pytest-django==3.9.0
pytest-dotenv==0.5.2
pytest-env==0.6.2
python-decouple==3.4
python-dotenv==0.15.0
pytz==2020.5
redis==3.5.3
regex==2020.11.13
requests==2.25.1
rfc3986==1.4.0
scout-apm==2.18.0
sentry-sdk==0.19.5
service-identity==18.1.0
six==1.15.0
sniffio==1.2.0
sqlparse==0.4.1
toml==0.10.2
Twisted==20.3.0
txaio==20.12.1
typed-ast==1.4.2
typing-extensions==3.7.4.3
urllib3==1.26.2
vine==5.0.0
wcwidth==0.2.5
whitenoise==5.2.0
wrapt==1.12.1
zipp==3.4.0
zope.interface==5.2.0
